css modal, fecha de ultima edicion en nota

+++ labnotes/src/components/Note.js
@@ -4,7 +4,9 @@ import { doc, deleteDoc } from "firebase/firestore";
 import { Modal } from './Modal';
 
 export const Note = ({ note }) => {
-	const { id, nota, titulo } = note;
+	 const fecha = new Date(note.fecha).toLocaleDateString('es-MX',{year:'numeric', month:'2-digit', day:'2-digit'});
+	console.log(fecha);
+	const { id, nota, titulo} = note;
 
 	const [isVisible, setIsVisible] = useState(false);
 
@@ -12,34 +14,27 @@ export const Note = ({ note }) => {
  const hideModal = () => setIsVisible(false);
 
 	const deleteNote = async () => {
-		// try {
-		// 	let opcion = confirm('Clicka en Aceptar o Cancelar');
-		// 			if (opcion === true) {
-		// 				await deleteDoc(doc(firebaseDB, 'notes', id));
-		// 	} else {
-		// 			alert ('elegiste eliminar');
-		// 	}}
-		// 	catch (error) {
-		// 			console.error(error);
-		// 		}
 		try {
-			alert ('Deseas eliminar la nota')
+			if(window.confirm('Deseas eliminar la nota')){
 				await deleteDoc(doc(firebaseDB, 'notes', id));
+			}
 			}catch (error) {
 			console.error(error);
 		}
 	}
 
 	return (
-		<article className='note'>
-			<div className='note-header'>
+		<article className='note-card'>
+			<div className='note-card-header'>
 				<h3>{titulo}</h3>
 			</div>
-				<div className='note-body'>
-				<p>{nota}</p>
+				<div className='note-card-body'>
+					<p>{nota}</p>
+					<br/>
 				</div>
-			<div className='note-footer'>
-				<div className="action-icons float-right">
+			<div className='note-card-footer'>
+				<p className='date-last-modification'>{fecha}</p>
+				<div className="action-icons">
 				<i className="material-icons action-icon" role="button" title="Edit" onClick={showModal}>edit</i>
 				<i className="material-icons action-icon" role="button" title="Delete" onClick={deleteNote} >delete</i>
 				</div>
